feat: pagination for ingredient search in storage

+++ src/handlers/storage/ingredients/view.cpp
@@ -6,6 +6,7 @@
 #include "message_tracker.hpp"
 #include "render/storage/ingredients/view.hpp"
 #include "render/storage/view.hpp"
+#include "tg_types.hpp"
 #include "utils/parsing.hpp"
 
 #include <algorithm>
@@ -19,6 +20,28 @@ using namespace render::storage;
 using namespace render::storage::ingredients;
 using namespace api::models::ingredient;
 
+// Global vars
+const size_t numOfIngredientsOnPage = 5;
+const size_t threshhold = 50;
+
+namespace {
+void updateSearch(StorageIngredientsList& state, BotRef bot, tg_types::UserId userId, IngredientsApiRef api) {
+
+    auto response = api.searchForStorage(
+        userId, state.storageId, "", threshhold, numOfIngredientsOnPage, state.pageNo * numOfIngredientsOnPage);
+    if (response.found != state.totalFound || !std::ranges::equal(response.page,
+                                                                  state.searchItems,
+                                                                  std::ranges::equal_to{},
+                                                                  &IngredientSearchForStorageItem::id,
+                                                                  &IngredientSearchForStorageItem::id)) {
+        state.searchItems = std::move(response.page);
+        state.totalFound = response.found;
+        if (auto mMessageId = message::getMessageId(userId))
+            renderIngredientsListSearch(state, numOfIngredientsOnPage, userId, userId, bot);
+    }
+}
+} // namespace
+
 void handleStorageIngredientsListCQ(
     StorageIngredientsList& state, CallbackQueryRef cq, BotRef bot, SMRef stateManager, ApiClientRef api) {
     bot.answerCallbackQuery(cq.id);
@@ -30,47 +53,52 @@ void handleStorageIngredientsListCQ(
         stateManager.put(StorageView{state.storageId});
         return;
     }
-
-    auto mIngredient = utils::parseSafe<api::IngredientId>(cq.data);
-    if (!mIngredient)
+    if (cq.data == "prev") {
+        state.pageNo -= 1;
+        updateSearch(state, bot, userId, api);
         return;
-    auto it = std::ranges::find(state.searchItems, *mIngredient, &IngredientSearchForStorageItem::id);
-    if (it == state.searchItems.end())
-        return;
-
-    if (it->available) {
-        api.getIngredientsApi().deleteFromStorage(userId, state.storageId, *mIngredient);
-        state.storageIngredients.remove(*mIngredient);
-    } else {
-        api.getIngredientsApi().putToStorage(userId, state.storageId, *mIngredient);
-        state.storageIngredients.put({.id = it->id, .name = it->name});
     }
-    it->available = !it->available;
 
-    renderIngredientsListSearch(state, userId, chatId, bot);
+    if (cq.data == "next") {
+        state.pageNo += 1;
+        updateSearch(state, bot, userId, api);
+        return;
+    }
+
+    if (cq.data != "dont_handle") {
+
+        auto mIngredient = utils::parseSafe<api::IngredientId>(cq.data);
+        if (!mIngredient)
+            return;
+        auto it = std::ranges::find(state.searchItems, *mIngredient, &IngredientSearchForStorageItem::id);
+        if (it == state.searchItems.end())
+            return;
+
+        if (it->available) {
+            api.getIngredientsApi().deleteFromStorage(userId, state.storageId, *mIngredient);
+            state.storageIngredients.remove(*mIngredient);
+        } else {
+            api.getIngredientsApi().putToStorage(userId, state.storageId, *mIngredient);
+            state.storageIngredients.put({.id = it->id, .name = it->name});
+        }
+        it->available = !it->available;
+        renderIngredientsListSearch(state, numOfIngredientsOnPage, userId, chatId, bot);
+    }
 }
 
 void handleStorageIngredientsListIQ(StorageIngredientsList& state,
                                     InlineQueryRef iq,
                                     BotRef bot,
                                     IngredientsApiRef api) {
+    const size_t numOfIngredientsOnPage = 5;
     const auto userId = iq.from->id;
     if (iq.query.empty()) {
         state.searchItems.clear();
-        renderIngredientsListSearch(state, userId, userId, bot);
+        // When query is empty then search shouldn't happen
+        state.totalFound = 0;
+        renderIngredientsListSearch(state, numOfIngredientsOnPage, userId, userId, bot);
     } else {
-        const std::size_t count = 100;
-        auto response = api.searchForStorage(userId, state.storageId, iq.query, count, 0);
-        if (response.found != state.totalFound || !std::ranges::equal(response.page,
-                                                                      state.searchItems,
-                                                                      std::ranges::equal_to{},
-                                                                      &IngredientSearchForStorageItem::id,
-                                                                      &IngredientSearchForStorageItem::id)) {
-            state.searchItems = std::move(response.page);
-            state.totalFound = response.found;
-            if (auto mMessageId = message::getMessageId(userId))
-                renderIngredientsListSearch(state, userId, userId, bot);
-        }
+        updateSearch(state, bot, userId, api);
     }
     // Cache is not disabled on Windows and Linux desktops. Works on Android and Web
     bot.answerInlineQuery(iq.id, {}, 0);
